sdk/python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moondb"
version = "0.1.0"
description = "Python SDK for MoonDB — Redis-compatible server with vector search and graph engine"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Tin Dang"},
]
keywords = ["moondb", "redis", "vector-search", "graph", "database"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "redis>=5.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.3"]
llamaindex = ["llama-index-core>=0.11"]
numpy = ["numpy>=1.24"]
all = ["moondb[langchain]", "moondb[llamaindex]", "moondb[numpy]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "mypy>=1.0",
    "ruff>=0.4",
]

[project.urls]
Homepage = "https://github.com/tindang/moon"
Documentation = "https://github.com/tindang/moon/tree/main/sdk/python"
Repository = "https://github.com/tindang/moon"

[tool.setuptools.packages.find]
include = ["moondb*"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "ANN", "B", "SIM"]
ignore = ["ANN101", "ANN102"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
